루틴 생성, readline으로 입력받은 후 실행
