+++ CodeStarterCamp_Week4/bodyCondition.swift
@@ -54,38 +54,3 @@ class BodyCondition {
         )
     }
 }
-
-struct Activity {
-    let name: String
-    let action: (BodyCondition) -> Void
-}
-
-let 윗몸일으키기: Activity = Activity(name: "윗몸일으키기") { bodyCondition in
-    print("<<\(윗몸일으키기.name)을(를) 시작합니다.>>")
-    bodyCondition.upperBodyStrength += Int.random(in: 10...20)
-    bodyCondition.fatigability += Int.random(in: 10...20)
-    bodyCondition.conditionCheck()
-}
-
-let 스쿼트: Activity = Activity(name: "스쿼트") { bodyCondition in
-    print("<<\(스쿼트.name)을(를) 시작합니다.>>")
-    bodyCondition.lowerBodyStrength += Int.random(in: 20...30)
-    bodyCondition.fatigability += Int.random(in: 10...20)
-    bodyCondition.conditionCheck()
-}
-
-let 오래달리기: Activity = Activity(name: "오래달리기") { bodyCondition in
-    print("<<\(오래달리기.name)을(를) 시작합니다.>>")
-    bodyCondition.upperBodyStrength += Int.random(in: 5...10)
-    bodyCondition.lowerBodyStrength += Int.random(in: 5...10)
-    bodyCondition.muscularEndurance += Int.random(in: 20...30)
-    bodyCondition.fatigability += Int.random(in: 20...30)
-    bodyCondition.conditionCheck()
-}
-
-let 동적휴식: Activity = Activity(name: "동적휴식") { bodyCondition in
-    print("<<\(동적휴식.name)을(를) 시작합니다.>>")
-    bodyCondition.fatigability -= Int.random(in: 5...10)
-    bodyCondition.conditionCheck()
-}
-
